fix(react): moved PlayerEvents to common package
