+++ packages/react-player/src/index.ts
@@ -2,12 +2,12 @@
  * Copyright 2023 Design Barn Inc.
  */
 
-import { PlayMode, PlayerState } from 'common';
+import { PlayMode, PlayerState, PlayerEvents } from 'common';
 import type { Manifest, ManifestAnimation } from 'common';
 
 import type { DotLottieRefProps } from './hooks/use-dotlottie-player';
 
-export { PlayMode, PlayerState, DotLottieRefProps };
+export { PlayerEvents, PlayMode, PlayerState, DotLottieRefProps };
 export type { Manifest, ManifestAnimation };
 export * from './react-player';
 export * from './controls';
