+++ packages/react-player/src/react-player.tsx
@@ -3,7 +3,7 @@
  */
 
 import type { RendererSettings, PlayMode } from 'common';
-import { PlayerState } from 'common';
+import { PlayerState, PlayerEvents } from 'common';
 import React, { useEffect, useRef } from 'react';
 import type { MutableRefObject } from 'react';
 
@@ -11,20 +11,6 @@ import { DotLottieContext } from './dotlottie-context';
 import type { DotLottieRefProps } from './hooks/use-dotlottie-player';
 import { useDotLottiePlayer } from './hooks/use-dotlottie-player';
 import { useSelectDotLottieState } from './hooks/use-select-dotlottie-state';
-
-export enum PlayerEvents {
-  Complete = 'complete',
-  DataFail = 'data_fail',
-  DataReady = 'data_ready',
-  Error = 'error',
-  Frame = 'frame',
-  Freeze = 'freeze',
-  LoopComplete = 'loopComplete',
-  Pause = 'pause',
-  Play = 'play',
-  Ready = 'ready',
-  Stop = 'stop',
-}
 
 export interface DotLottiePlayerProps extends React.HTMLAttributes<HTMLDivElement> {
   activeAnimationId?: string;
